Move file bar particulars
